requirements.txt
aiocache==0.12.2
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
dispie==0.0.11
dnspython==2.6.1
easy-pil==0.3.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
multidict==6.0.5
mysql-connector-python==8.3.0
packaging==23.2
Pillow==10.1.0
pydantic==2.6.1
pydantic_core==2.16.2
python-engineio==4.9.0
python-socketio==5.4.1
requests==2.31.0
websockets
simple-websocket==1.0.0
sniffio==1.3.0
urllib3==2.2.1
uvicorn==0.27.0.post1
wsproto==1.2.0
yarl==1.9.4
redis==5.0.3

PyNaCl==1.5.0
fastapi==0.109.2
starlette==0.36.3
authlib==1.3.0
cffi==1.16.0
cryptography==42.0.5
pycparser==2.22
aiocache==0.12.2
Jinja2==3.1.3
docker==7.0.0

better-ipc==2.0.3
# discord-ext-ipcx==0.2.2
# better-ipc
# discord-ext-ipc==2.1.1
# typing_extensions==4.9.0